requirements.txt
Flask_WTF==0.14.2
Keras==2.2.4
numpy==1.14.2
Flask==1.0.2
WTForms==2.1
Flask_Bootstrap==3.3.7.1
Pillow==6.0.0
